Use imagemagick for HEIC image conversion on Linux
