+++ go.mod
@@ -11,7 +11,6 @@ require (
 	github.com/jessevdk/go-flags v1.6.1
 	github.com/mattn/go-sqlite3 v1.14.42
 	github.com/spf13/afero v1.15.0
-	github.com/tagatac/goheif v0.0.0-20260419130429-f87270e9cc43
 	github.com/tagatac/gorecurcopy v1.0.1
 	github.com/tagatac/typedstream-go v0.0.0-20260419125731-cd361d54cae6
 	go.uber.org/mock v0.6.0
@@ -20,7 +19,6 @@ require (
 
 require (
 	github.com/google/go-cmp v0.7.0 // indirect
-	github.com/pkg/errors v0.9.1 // indirect
 	golang.org/x/sys v0.43.0 // indirect
 	golang.org/x/text v0.36.0 // indirect
 )
