+++ imgconv/convert_darwin.go
@@ -2,25 +2,12 @@ package imgconv
 
 import "os/exec"
 
-type imgConverter struct {
-	execCommand func(string, ...string) *exec.Cmd
-	tempDir     string
-}
-
-func NewImgConverter(tempDir string) ImgConverter {
-	return &imgConverter{
-		execCommand: exec.Command,
-		tempDir:     tempDir,
-	}
-}
-
-func (i *imgConverter) convert(src, dst string) error {
-	cmd := i.execCommand(
+func (i *imgConverter) convCmd(src, dst string) *exec.Cmd {
+	return i.execCommand(
 		"sips",
 		"--setProperty", "format", "jpeg",
 		"--setProperty", "formatOptions", "best",
 		"--out", dst,
 		src,
 	)
-	return cmd.Run()
 }
