+++ imgconv/convert_linux.go
@@ -1,19 +1,7 @@
 package imgconv
 
-import "github.com/tagatac/goheif/heic2jpg"
+import "os/exec"
 
-type imgConverter struct {
-	heic2jpg.Converter
-	tempDir string
-}
-
-func NewImgConverter(tempDir string) ImgConverter {
-	return &imgConverter{
-		Converter: heic2jpg.NewConverter(heic2jpg.WithQuality(100)),
-		tempDir:   tempDir,
-	}
-}
-
-func (i *imgConverter) convert(src, dst string) error {
-	return i.HEIC2JPG(src, dst)
+func (i *imgConverter) convCmd(src, dst string) *exec.Cmd {
+	return i.execCommand("magick", "convert", src, dst)
 }
